+++ novae/__init__.py
@@ -11,6 +11,7 @@ from . import data
 from . import monitor
 from . import plot
 from .utils import spatial_neighbors, batch_effect_correction
+from .data import load_dataset
 
 log = logging.getLogger("novae")
 configure_logger(log)