+++ novae/data/__init__.py
@@ -1,3 +1,7 @@
 from .convert import AnnDataTorch
 from .dataset import NovaeDataset
 from .datamodule import NovaeDatamodule
+from ._embeddings._histo import compute_histo_embeddings, compute_histo_pca
+from .load import _load_wandb_artifact, load_dataset
+from .preprocess import quantile_scaling
+from .toy import toy_dataset