+++ novae/utils/__init__.py
@@ -1,22 +1,19 @@
 from ._utils import (
-    repository_root,
-    tqdm,
     fill_invalid_indices,
+    iter_slides,
     lower_var_names,
-    wandb_log_dir,
-    pretty_num_parameters,
-    pretty_model_repr,
     parse_device_args,
-    valid_indices,
+    pretty_model_repr,
+    pretty_num_parameters,
+    repository_root,
+    sparse_std,
+    tqdm,
     unique_leaves_indices,
     unique_obs,
-    sparse_std,
-    iter_slides,
+    valid_indices,
+    wandb_log_dir,
 )
-from ._build import spatial_neighbors
-from ._validate import check_available_domains_key, prepare_adatas, check_has_spatial_adjancency, check_slide_name_key
-from ._data import load_local_dataset, load_wandb_artifact, toy_dataset, load_dataset
-from ._mode import Mode
-from ._correct import batch_effect_correction
-from ._preprocess import quantile_scaling
-from ._histo import compute_histo_embeddings, compute_histo_pca
+from ._validate import check_available_domains_key, check_has_spatial_adjancency, check_slide_name_key, prepare_adatas
+from .build import spatial_neighbors
+from .correct import batch_effect_correction
+from .mode import Mode