+++ tests/test_plots.py
@@ -3,7 +3,7 @@ import pytest
 import novae
 from novae._constants import Keys
 
-adatas = novae.utils.toy_dataset(n_panels=2, xmax=200)
+adatas = novae.data.toy_dataset(n_panels=2, xmax=200)
 novae.utils.spatial_neighbors(adatas)
 
 model = novae.Novae(adatas)